requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-swagger==2.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
drf-yasg==1.21.7
environ==1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
psycopg==3.1.18
psycopg2==2.9.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.1.1
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
